DataFinder-AI/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
sqlalchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
sentence-transformers>=3.0,<4.0
scikit-learn>=1.5,<2.0
requests>=2.32,<3.0
pytest>=8.0,<9.0
psycopg2-binary>=2.9,<3.0
faiss-cpu>=1.8,<2.0
apscheduler>=3.10,<4.0
huggingface_hub>=0.24,<1.0
kaggle>=1.6,<2.0
ucimlrepo>=0.0.7,<1.0
